// src/app/modules/openapi/model/token.ts
/**
 * E-Commerce Application
 *
 * Contact: sad@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Token { 
    userId?: string;
    accessToken?: string;
    refreshToken?: string;
}